backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
asyncpg==0.29.0
langchain==0.1.16
langchain-groq==0.1.0
langchain-huggingface==0.0.1
langchain-community==0.0.34
faiss-cpu==1.7.4
pypdf==3.17.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.25.2
sentence-transformers==2.6.1
